+++ changelog.md
@@ -1,7 +1,11 @@
 
 ## To add
 
+
 ## Added
+* Trees
+* User Control for camera rotation
+* Sea
 * Rotating camera
 * Different terrain types
 * Procedural generation of terrain